+++ components/org.wso2.carbon.identity.sso.agent/src/main/java/org/wso2/carbon/identity/sso/agent/util/SSOAgentRequestResolver.java
@@ -18,7 +18,7 @@
  *
  */
 
-package org.wso2.carbon.identity.sso.agent;
+package org.wso2.carbon.identity.sso.agent.util;
 
 import org.opensaml.common.xml.SAMLConstants;
 import org.wso2.carbon.identity.sso.agent.bean.LoggedInSessionBean;
